Back to normal + icon
i delete that Gui folder and i have put hte gui.py,  where it was before and add an icon.ico that needs for the gui.py to work
